Restructure test/benchmarks
