+++ reasoner/BUILD
@@ -26,8 +26,8 @@ package(default_visibility = [
     "//server:__pkg__",
     "//test/integration/reasoner:__pkg__",
     "//test/behaviour/reasoner:__subpackages__",
-    "//test/benchmark/reasoner:__pkg__",
-    "//test/benchmark/iam:__pkg__",
+    "//test/benchmark/reasoner/iam:__subpackages__",
+    "//test/benchmark/reasoner/synthetic:__subpackages__"
 ])
 
 native_java_libraries(
