+++ test/benchmark/reasoner/iam/ConjunctionStructureTest.java
