+++ test/benchmark/reasoner/iam/resources/data_large.typedb
