+++ test/benchmark/reasoner/iam/resources/schema_rules_optimised.tql
