+++ test/benchmark/reasoner/synthetic/BenchmarkBig.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.benchmark;
+package com.vaticle.typedb.core.reasoner.benchmark.synthetic;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
@@ -81,7 +81,7 @@ public class BenchmarkBig {
     final private Random rand = new Random();
 
     private void loadSchema(String fileName, TypeDB.Session session) {
-        String filePath = "test/benchmark/reasoner/resources/" + fileName;
+        String filePath = "test/benchmark/reasoner/synthetic/resources/" + fileName;
         try (TypeDB.Transaction tx = session.transaction(Arguments.Transaction.Type.WRITE)) {
             tx.query().define(Util.parseTQL(filePath).asDefine());
             tx.commit();
