+++ test/benchmark/reasoner/synthetic/generation/PathTreeGraph.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.test.benchmark.generation;
+package com.vaticle.typedb.core.reasoner.benchmark.synthetic.generation;
 
 import com.vaticle.typedb.core.TypeDB;
 import com.vaticle.typedb.core.common.parameters.Arguments;
@@ -25,7 +25,7 @@ import com.vaticle.typedb.core.concept.thing.Relation;
 import com.vaticle.typedb.core.concept.thing.Thing;
 import com.vaticle.typedb.core.concept.type.RelationType;
 import com.vaticle.typedb.core.concept.type.RoleType;
-import com.vaticle.typedb.core.reasoner.benchmark.Util;
+import com.vaticle.typedb.core.reasoner.benchmark.synthetic.Util;
 
 /**
  * Defines a Graph based on test 6.10 from Cao p. 82.
@@ -35,7 +35,7 @@ public class PathTreeGraph {
     private final TypeDB.DatabaseManager databaseManager;
     private final String databaseName;
 
-    private final static String schemaPath = "test/benchmark/reasoner/resources/";
+    private final static String schemaPath = "test/benchmark/reasoner/synthetic/resources/";
     private final String schemaFile;
     private final static Label key = Label.of("index");
