+++ test/benchmark/reasoner/synthetic/resources/diagonalTest.tql
