+++ test/benchmark/reasoner/synthetic/resources/linearTransitivity.tql
