+++ test/benchmark/reasoner/synthetic/resources/multiJoin.tql
