+++ test/benchmark/reasoner/synthetic/resources/pathTest.tql
