+++ test/benchmark/reasoner/synthetic/resources/quadraticTransitivity.tql
